Create LICENSE
I chose GPL v3 for this project because the original game itself is public domain. So any derivations of that original should be free to anyone for any purpose. The original source code is lost to time, and was not used for this project. So the source code is unique and original, even if the game itself is not. But I think this captures the spirit of the game's place in history best.